+++ e2e/runner/sui.go
@@ -93,7 +93,7 @@ func (r *E2ERunner) SuiWithdrawAndCallSUI(
 		payloadBytes,
 		gatewayzevm.CallOptions{
 			IsArbitraryCall: false,
-			GasLimit:        big.NewInt(25000),
+			GasLimit:        big.NewInt(100000),
 		},
 		revertOptions,
 	)
@@ -143,7 +143,7 @@ func (r *E2ERunner) SuiWithdrawAndCallFungibleToken(
 		payloadBytes,
 		gatewayzevm.CallOptions{
 			IsArbitraryCall: false,
-			GasLimit:        big.NewInt(20000),
+			GasLimit:        big.NewInt(100000),
 		},
 		revertOptions,
 	)